Fixed all the jUnit test failures.  No new feature tests.
